+++ packages/shadergradient-v2/src/ShaderGradient/Mesh/Mesh.tsx
@@ -9,6 +9,7 @@ import {
 } from '@/shaders/a'
 // import { vertexShader, fragmentShader } from '@/shaders/base'
 import { dToRArr } from '@/utils'
+import { debug } from '@/utils/debug'
 
 export function Mesh({
   animate,
@@ -56,7 +57,7 @@ export function Mesh({
         vertexShader={type === 'sphere' ? vertexSphere : vertexShader}
         fragmentShader={type === 'sphere' ? fragmentSphere : fragmentShader}
         onInit={(material) => {
-          // console.log('material (onInit)', material)
+          debug.performance('material (onInit)', material);
         }}
       />
     </mesh>