+++ packages/ui/src/components/Shared/ShaderGradientStateless/ShaderGradientStatelessCanvas.tsx
@@ -1,0 +1,4 @@
+import { ShaderGradientCanvas } from '@shadergradient/react'
+
+const ShaderGradientStatelessCanvas = ShaderGradientCanvas
+export { ShaderGradientStatelessCanvas }