+++ packages/ui/src/components/Web/index.ts
@@ -1,0 +1,8 @@
+export * from './Button'
+
+export * from './Icons'
+export * from './SimpleSlider'
+
+export * from './ScrollableBox'
+
+export * from './TriggerSection'