+++ packages/ui/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Shared'
+export * from './Web'
+export * from './FigmaPlugin'