+++ packages/ui/src/index.ts
@@ -1,8 +1,4 @@
 export * from './store'
 
-export * from './Button'
-export * from './TextAnimation'
-export * from './TextHover'
-
-export * from './PluginUI'
-export * from './WebUI'
+export * from './components'
+export * from './overrides'